+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # CHANGELOG
 
+## Version 1.3.0 - TBD
+
+### Add - TODO
+* Site footer
+    * Version badge (add to footer) ![](https://img.shields.io/github/v/tag/meditativesphynx/portfolio)
+    * Link to GitHub repo for site (add to footer) ![](https://img.shields.io/badge/GitHub-Repo-blue?link=https://github.com/MeditativeSphynx/portfolio&logo=github)
+* Home page panels to showcase skills learned
+
+### Fix - TODO
+* Mobile wordwrap on home page
+* Top-nav icon href to '/' instead of 'home'
+
+
+## Version 1.2.0 - May 17, 2023
+
+### Added
+* Site changelog page
+* Changelog to topnav
+* My photo to the welcome page
+
 ## Version 1.1.3
 
 ### Fixed